Rename DTO namespace to ValueObject
